requirements.txt
# Core LangChain and ecosystem
langchain
langchain-community==0.3.23
langchain-core==0.3.56
langchain-text-splitters==0.3.8
langsmith==0.3.37
langchain-openai

# Vector DB
pinecone==6.0.2
pinecone-client==2.2.1
pinecone-plugin-interface==0.0.7

# LLM Provider
groq==0.23.1

# FastAPI stack
fastapi
starlette==0.36.3  # Compatible with fastapi 0.110.0
uvicorn

# Pydantic (match with langchain and fastapi)
pydantic
pydantic-core==2.33.1
pydantic-settings==2.9.1

# Utilities and helpers
python-dotenv==0.21.1
requests==2.28.1
aiohttp==3.11.18
httpx==0.28.1
httpx-sse==0.4.0
anyio==4.9.0
tenacity==8.5.0
tqdm==4.67.1
loguru==0.7.3
typing-inspect==0.9.0
typing-extensions==4.13.2
colorama==0.4.6
structlog==21.3.0

# PDF/document parsing
pdfminer.six==20221105
PyMuPDF==1.25.5

# Serialization
dataclasses-json==0.5.14
orjson==3.10.16
PyYAML==6.0.2

# Parsing
marshmallow==3.26.1

# Date/time
python-dateutil==2.9.0.post0

# System packages
certifi==2025.4.26
charset-normalizer==2.1.1
idna==3.10
urllib3==1.26.20
zstandard==0.23.0
cffi==1.17.1
cryptography==44.0.2
pycparser==2.22
greenlet==3.2.1
SQLAlchemy==1.4.54

# Optional
click==8.1.8
distro==1.9.0
sniffio==1.3.1
win32-setctime==1.2.0
streamlit==1.34.0
langchain==0.1.17
langchain-community==0.0.36
groq==0.5.0
openai==1.30.1
pinecone-client==3.1.0
python-dotenv==1.0.1
tiktoken==0.7.0
PyPDF2==3.0.1
docx2txt==0.8
structlog